// server/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  admin: 'admin',
  user: 'user'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const KycStatus = {
  pending: 'pending',
  verified: 'verified',
  rejected: 'rejected',
  resubmit: 'resubmit'
} as const

export type KycStatus = (typeof KycStatus)[keyof typeof KycStatus]


export const GovernmentIdType = {
  international_passport: 'international_passport',
  national_id: 'national_id',
  driving_license: 'driving_license'
} as const

export type GovernmentIdType = (typeof GovernmentIdType)[keyof typeof GovernmentIdType]


export const AccountType = {
  personal: 'personal',
  business: 'business'
} as const

export type AccountType = (typeof AccountType)[keyof typeof AccountType]


export const AccountOwnership = {
  single: 'single',
  joint: 'joint'
} as const

export type AccountOwnership = (typeof AccountOwnership)[keyof typeof AccountOwnership]


export const AccountStatus = {
  active: 'active',
  dormant: 'dormant',
  closed: 'closed'
} as const

export type AccountStatus = (typeof AccountStatus)[keyof typeof AccountStatus]


export const AccountUserRole = {
  owner: 'owner',
  co_owner: 'co_owner',
  manager: 'manager',
  admin: 'admin',
  accountant: 'accountant',
  investor: 'investor',
  contributor: 'contributor',
  legal_guardian: 'legal_guardian',
  signatory: 'signatory'
} as const

export type AccountUserRole = (typeof AccountUserRole)[keyof typeof AccountUserRole]


export const JointAccountRequestStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected'
} as const

export type JointAccountRequestStatus = (typeof JointAccountRequestStatus)[keyof typeof JointAccountRequestStatus]


export const JointAccountModRequestType = {
  transfer: 'transfer',
  withdrawal: 'withdrawal',
  name_change: 'name_change'
} as const

export type JointAccountModRequestType = (typeof JointAccountModRequestType)[keyof typeof JointAccountModRequestType]


export const ProfitDistribution = {
  daily: 'daily',
  weekly: 'weekly',
  bi_weekly: 'bi_weekly',
  monthly: 'monthly'
} as const

export type ProfitDistribution = (typeof ProfitDistribution)[keyof typeof ProfitDistribution]


export const InvestmentPlanCategory = {
  forex: 'forex',
  stocks: 'stocks',
  real_estate: 'real_estate',
  bonds: 'bonds',
  commodities: 'commodities',
  cryptocurrencies: 'cryptocurrencies',
  derivatives: 'derivatives'
} as const

export type InvestmentPlanCategory = (typeof InvestmentPlanCategory)[keyof typeof InvestmentPlanCategory]


export const InvestmentStatus = {
  open: 'open',
  closed: 'closed',
  paused: 'paused',
  terminated: 'terminated'
} as const

export type InvestmentStatus = (typeof InvestmentStatus)[keyof typeof InvestmentStatus]


export const TransactionType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
  transfer: 'transfer',
  investment: 'investment',
  profit: 'profit'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const TransactionStatus = {
  pending: 'pending',
  successfull: 'successfull',
  reversed: 'reversed',
  failed: 'failed'
} as const

export type TransactionStatus = (typeof TransactionStatus)[keyof typeof TransactionStatus]


export const TransactionMedium = {
  wire: 'wire',
  crypto: 'crypto'
} as const

export type TransactionMedium = (typeof TransactionMedium)[keyof typeof TransactionMedium]


export const NotificationBodyType = {
  string: 'string',
  html: 'html'
} as const

export type NotificationBodyType = (typeof NotificationBodyType)[keyof typeof NotificationBodyType]